backend/requirements.txt
django
djangorestframework
django-cors-headers
django-storages
boto3
python-dotenv